Moved ByteStringInternal back
Ignore-this: 3a8ff821a5b4c2627a89926d51dfabd8

darcs-hash:1ba249ac32efcf25568a61288bf5d05c59b57201
