+++ src/Data/Trie/ByteStringInternal/indexOfDifference.c
@@ -1,7 +1,7 @@
 /* -------------------------------------------------------------
 --                                                  ~ 2009.01.07
 -- |
--- Module      :  Data.Trie.BigEndianPatricia.ByteStringInternal.indexOfDifference
+-- Module      :  Data.Trie.ByteStringInternal.indexOfDifference
 -- Copyright   :  Copyright (c) 2008--2011 wren gayle romano
 -- License     :  BSD3
 -- Maintainer  :  wren@community.haskell.org
